% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltration.R
\name{infiltration}
\alias{infiltration}
\title{Cumulative infiltration and sqrt of time}
\usage{
infiltration(dataset, col_name, ...)
}
\value{
A tibble giving three new columns: sqrt_time,
volume_infiltrated and infiltration
}
\description{
Using
}
\examples{
infiltration(tibble(time = c(0, 30, 60, 90, 120, 150, 180, 210, 240, 270, 300,
0, 35, 65, 95, 125, 155, 185, 215, 245, 275, 305),
volume = c(95, 89, 86, 83, 80, 77, 74, 73, 71, 69, 67,
           83, 77, 64, 61, 58, 45, 42, 35, 29, 17, 15)))
}
