% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_A.R
\name{parameter_A}
\alias{parameter_A}
\title{Calculates parameter A from (Philip, 1957)}
\usage{
parameter_A(dataset, col_name, ...)
}
\arguments{
\item{dataset}{tibble or dataset}

\item{col_name}{including suction, n_ho, alpha and suction}
}
\value{
A tibble giving two new columns: suction_num,
and parameter_A
}
\description{
Calculates parameter A from (Philip, 1957)
}
\examples{
parameter_A(tibble(alpha = c(0.145, 0.008), n_ho = c(2.68, 1.09),suction = c("2cm","3cm")))
}
