% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg_par.R
\name{vg_par}
\alias{vg_par}
\title{Tabulated VG parameters
Van Genuchten parameters and values of A, n and alpha for the Minidisk
Infiltrometer (Decagon Devices, Inc., 2005).
12 soil texture classes and suction from
0.5 to 7 cm are tabulated}
\usage{
vg_par(dataset, col_name)
}
\arguments{
\item{dataset}{A \code{\link[=tibble]{tibble()}} or \code{\link[=data.frame]{data.frame()}} including suction and texture}

\item{col_name}{\code{\link[=vars]{vars()}} including suction and texture}
}
\value{
A tibble giving three new columns: n_ho,
alpha and A value
}
\description{
Tabulated VG parameters
Van Genuchten parameters and values of A, n and alpha for the Minidisk
Infiltrometer (Decagon Devices, Inc., 2005).
12 soil texture classes and suction from
0.5 to 7 cm are tabulated
}
\examples{
vg_par(data.frame(suction = c("2cm", "3cm"), texture = c("sand", "clay")))
}
