% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_A.R
\name{parameter_A}
\alias{parameter_A}
\title{Calculates parameter A from (Philip, 1957)}
\usage{
parameter_A(dataset, col_name)
}
\arguments{
\item{dataset}{A \code{\link[=tibble]{tibble()}} or \code{\link[=data.frame]{data.frame()}} including n_ho, alpha and suction}

\item{col_name}{\code{\link[=vars]{vars()}} including n_ho, alpha and suction}
}
\value{
A \code{\link[=tibble]{tibble()}} giving two new columns: suction_num,
and parameter_A
}
\description{
Calculates parameter A from (Philip, 1957)
}
\examples{
parameter_A(data.frame(alpha = c(0.145, 0.008), n_ho = c(2.68, 1.09),suction = c("2cm","3cm")))
}
