% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg_par.R
\name{vg_par}
\alias{vg_par}
\title{Tabulated VG parameters
Van Genuchten parameters and values of A, n and alpha for the Minidisk
Infiltrometer (Decagon Devices, Inc., 2005).
12 soil texture classes and suction from
0.5 to 7 cm are tabulated}
\usage{
vg_par(suction, texture, ...)
}
\arguments{
\item{suction}{a chr or vector of chr}

\item{texture}{a chr or vector of chr}
}
\value{
A tibble giving three new columns: n_ho,
alpha and value
}
\description{
Tabulated VG parameters
Van Genuchten parameters and values of A, n and alpha for the Minidisk
Infiltrometer (Decagon Devices, Inc., 2005).
12 soil texture classes and suction from
0.5 to 7 cm are tabulated
}
\examples{
vg_par(suction = c("2cm", "3cm"), texture = c("sand", "clay"))
}
